+++ CHANGELOG.md
@@ -34,6 +34,8 @@ Notable changes to the `alacritty_terminal` crate are documented in its
 
 - Spurious "Failed to set new owner of XCB selection" warnings on X11
 - Lacking permissions to launch software sending Apple events
+- Built-in Braille dot mapping, additional legacy computing block rendering, and Powerline
+  Extra symbols
 
 ## 0.17.0